+++ storybook/stories/support/StoryDecorator.js
@@ -26,7 +26,10 @@ export default class StoryDecorator extends Component<Props> {
         <IntlProvider {...{ locale: 'en-US', key: 'en-US', messages: translations['en-US'] }}>
           <ThemeProvider theme={daedalusTheme}>
             <div>
-              {Children.map(children, (child) => React.cloneElement(child, { propsForChildren }))}
+              {Children.map(children, (child) => {
+                const childProps = child.type === 'div' ? {} : { propsForChildren };
+                return React.cloneElement(child, childProps);
+              })}
             </div>
           </ThemeProvider>
         </IntlProvider>