+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 ### Chores
 
 - Integrates the latest features from the develop branch of react-polymorph. Features include: render props architecture, theme composition, and a ThemeProvider HOC. [PR 950](https://github.com/input-output-hk/daedalus/pull/950)
+- Fixed broken storybook [PR 1041](https://github.com/input-output-hk/daedalus/pull/1041)
 
 ### Features